Create Questions
The average revenue per transaction in the population is Rs 614. Based on this attribute only, would you be confident that this sample is representative of the population?

Hint: Is the sample average different from the population average? If yes, how different based on statistical significance? Please note that sample size > 100
Type in the p-value of your hypothesis here, rounded to two decimals (please use the format 0.xx)

2. The average revenue per transaction in the population is Rs 614. Based on this attribute only, would you be confident that this sample is representative of the population?  

Hint: Is the sample average different from the population average? If yes, how different based on statistical significance? Please note that sample size > 100

Choose the right option for the text boxes in the sentence below:

Based on the p-value, I will ________________ and conclude that the sample is ___________________

3. The average order rate (Orders with revenue > 0) is 58% in the population. You will need to assess how likely is it that the observed sample order rate or greater is because of random chance.

Based on the p-value you find, should you continue working with this sample if order rate is a critical piece of your analysis? Explain how you reach your conclusion. 

Please note that this is a 5 point question so your explanation should be as detailed as possible.

 

Hint: The order rate is based on a variable that has 1/0 values. We want to compare the order rate of sample vs the order rate in the population and then to find out the likelihood of seeing the sample order rate simply because of randomness.

4. You are trying to assess if there is a significant difference in average time spent on the site between visitors that come in on Mondays only vs Sundays only

Run a hypothesis test and calculate the p-value. Please provide the p-value in the box below, rounded to two decimal places (in the format 0.xx)

5. What is the likelihood that based on this sample, a random customer will add atleast 3 products to his cart in a visit?

Insert the likelihood in percentage terms here (2 decimal places, in the format XX.xx) 

 

Hint: What is the average number of products in the cart in the sample?

6.One of the new members of your team has suggested that local offers and regional offers are one way of attracting more customers. Should you try to build a localization strategy? 

Please note that this is a business question that needs to be supported by data. Localization implies having different strategies for different areas, rather than having only one national strategy. 

This is a 5 point question, so please provide a detailed explanation of your conclusion and provide evidence to support your conclusion