Add files via upload
1. One of the new members of your team has suggested that local offers and regional offers are one way of attracting more customers. Should you try to build a localization strategy? 

Please note that this is a business question that needs to be supported by data. Localization implies having different strategies for different areas, rather than having only one national strategy. 

This is a 5 point question, so please provide a detailed explanation of your conclusion and provide evidence to support your conclusion

2.What is the likelihood that based on this sample, a random customer will add atleast 3 products to his cart in a visit?

Insert the likelihood in percentage terms here (2 decimal places, in the format XX.xx) 

 

Hint: What is the average number of products in the cart in the sample?